Added tests for reverseSentence. Tests are for reverseSentence with empty string, no words just punctuation, a single word and no punctuation, multiple words and no punctuation, and multiple words with punctuation.
Also changed reverseString from instance to static. I'm not sure that's
right, but I couldn't write static tests for instance methods, I got the
error "Cannot make a static reference to the non-static method
reverseSentence(String) from the type SentenceReverser" when calling it
in the actual value parameter of my tests. So I figured that I would
make the tests compile for now, as a start, and figure out whether
having the code static is a problem, and why I can't make the testing of
the instance methods compile, down the road.